parser/pyproject.toml
[project]
name = "parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncpg>=0.31.0",
    "cairosvg>=2.7.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
